properly clip spans against gspans and update gspan

+++ render/r_span.h
@@ -20,7 +20,7 @@ extern void
 R_SpanSetup (void);
 
 extern void
-R_EmitSpan (int y, int x1, int x2);
+R_EmitSpan (short y, short x1, short x2);
 
 extern void
 R_BeginSpanFrame (void *buf, int buflen);
